% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc2drug.R
\name{atc2drug}
\alias{atc2drug}
\title{Match ATC Codes to Drug Names}
\usage{
atc2drug(input_data, atc_column, reference_data)
}
\arguments{
\item{input_data}{A data frame containing a column of ATC codes to be matched.}

\item{atc_column}{The name of the column in \code{input_data} containing ATC codes.}

\item{reference_data}{A data frame with drug names in the first column and corresponding ATC codes in the second column.}
}
\value{
A modified data frame with an additional column:
\describe{
\item{atc_code_drug}{Drug name matched to each ATC code.}
}
}
\description{
This function matches ATC codes in the input data frame to corresponding drug
names using a reference data frame.
\cr
\cr
The reference data frame must have drug names in the first column and ATC codes in the second column.
}
\examples{
# Example usage:
data(example_atc_data)
data(example_reference_data)
atc2drug(example_atc_data, "ATC_code", example_reference_data)
}
