% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_atc_data}
\alias{example_atc_data}
\title{Example ATC Data for Testing}
\format{
A data frame with 8 rows and 1 variable:
\describe{
\item{ATC_code}{A character vector representing ATC codes, including valid codes, non-valid codes, and NA.}
}
}
\usage{
example_atc_data
}
\description{
This dataset provides an example of ATC codes for testing the \code{atc2drug} function.
It includes valid ATC codes, non-valid ATC codes, and NA values.
}
\examples{
example_atc_data
}
\keyword{datasets}
