% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_drug_data}
\alias{example_drug_data}
\title{Example Drug Data without Milligrams}
\format{
A data frame with 5 rows and 1 variable:
\describe{
\item{drug_name}{A character vector representing drug names.}
}
}
\usage{
example_drug_data
}
\description{
This dataset provides an example of drug names without milligrams.
}
\examples{
example_drug_data
}
\keyword{datasets}
