% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_drug_data_with_mg}
\alias{example_drug_data_with_mg}
\title{Example Drug Data with Milligrams}
\format{
A data frame with 5 rows and 1 variable:
\describe{
\item{drug_name}{A character vector representing drug names with milligrams.}
}
}
\usage{
example_drug_data_with_mg
}
\description{
This dataset provides an example of drug names including milligrams.
}
\examples{
example_drug_data_with_mg
}
\keyword{datasets}
