% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_reference_data}
\alias{example_reference_data}
\title{Example Reference Data for Drug Matching}
\format{
A list with 2 elements:
\describe{
\item{ATC_codes}{A character vector representing ATC codes.}
\item{drug_names}{A character vector representing corresponding drug names.}
}
}
\usage{
example_reference_data
}
\description{
This dataset provides an example reference data for drug matching.
}
\examples{
example_reference_data
}
\keyword{datasets}
